Add files via upload
The following is a full featured website which can be used as stock market calculator. The following website will show you all the stock related news plus a watchlist and a buy/sell engine giving a full featured info about the rising and sleeing stock prices,its best price,loss,gain,volume etc. The API used is alphavantage which provides the key for free, and there are several other apis as well